requirements.txt
anthropic==0.25.8
dspy_ai==2.4.5
langdetect==1.0.9
newspaper3k==0.2.8
nltk==3.5
openai==1.27.0
pandas==2.2.2
python-dotenv==1.0.1
rouge_score==0.1.2
scikit_learn==1.2.2
tqdm==4.66.4
